Fixing user, role and permission relations. Fixing security issue with principal. Adding user info to controllers, fixing busines rules.

+++ src/main/java/com/leandroinacio/picmeapi/user/UserService.java
@@ -7,8 +7,11 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+
+import com.leandroinacio.picmeapi.jwt.JwtUser;
 
 @Service
 public class UserService implements IUserService {
@@ -21,32 +24,63 @@ public class UserService implements IUserService {
 	@Autowired
 	private PasswordEncoder passwordEncoder;
 	
-	public void save(User user) {
-		if (user.getLastPasswordReset() == null) {
-			user.setLastPasswordReset(Calendar.getInstance());
+	public void save(User user, Authentication auth) {
+		
+		// Get the user if it's logged in
+		JwtUser jwtUser = null;
+		if (auth != null) {
+			jwtUser = (JwtUser) auth.getPrincipal();
 		}
-		user.setPassword(passwordEncoder.encode(user.getPassword()));
+		
+		// Return if logged used is trying to update a different user
+		if (jwtUser != null && user.getId() != jwtUser.getId()) {
+			return;
+		}
+		
+		// If it's new user, encode password and set last password reset
+		// Or if it's logged user changing password, do the same
+		if (jwtUser == null || (jwtUser != null && !user.getPassword().isEmpty())) {
+			user.setLastPasswordReset(Calendar.getInstance());
+			user.setPassword(passwordEncoder.encode(user.getPassword()));
+		}
+		
 		userRepository.save(user);
 	}
 	
 	public Page<User> findAll(Integer page, Integer size) {
-		return userRepository.findAll(PageRequest.of(page, size));
+		
+		// TODO: Change this so the query won't bring any password
+		Page<User> users = userRepository.findAll(PageRequest.of(page, size));
+		
+		users.getContent().forEach(user -> {
+			user.setPassword(null);
+		});
+		
+		return users;
 	}
 	
 	public User findById(Long id) {
-		return userRepository.findById(id);
+		User user = userRepository.findById(id);
+		user.setPassword(null);
+		return user;
 	}
 	
 	public User findByEmail(String email) {
-		return userRepository.findByEmail(email);
+		User user = userRepository.findByEmail(email);
+		user.setPassword(null);
+		return user;
 	}
 	
 	public User findByFirstName(String firstName) {
-		return userRepository.findByFirstName(firstName);
+		User user = userRepository.findByFirstName(firstName);
+		user.setPassword(null);
+		return user;
 	}
 	
-	public void deleteById(Long id) {
-		userRepository.deleteById(id);
+	public void deleteById(Long id, JwtUser jwtUser) {
+		if (id == jwtUser.getId()) {
+			userRepository.deleteById(id);
+		}
 	}	
 
 }
